backend/requirements.txt
asgiref==3.5.2
contourpy==1.0.5
cycler==0.11.0
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fonttools==4.37.3
fpdf==1.7.2
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2






